+++ CHANGES.md
@@ -21,7 +21,7 @@ Changes:
 -
 
 Fixes:
--
+- handle different table structures in `uscgcoca` scraper #1526
 
 ## Current